+++ src/components/GameSkeleton.jsx
