+++ src/components/Header.jsx
@@ -1,0 +1,11 @@
+import logo from '../assets/logo.png';
+
+const Header = () => {
+  return (
+    <header>
+      <img src={logo} alt="logo" className="logo" />
+    </header>
+  );
+};
+
+export default Header;
