+++ src/main.jsx
@@ -1,5 +1,5 @@
 import ReactDOM from 'react-dom/client';
 import App from './App.jsx';
-import './index.scss';
+import './styles/index.scss';
 
 ReactDOM.createRoot(document.getElementById('root')).render(<App />);
